Worked out weirdness with AJAX?
Still not quite sure if it works the way i expect...the put method in
the shapes view doesn't seem to be run, but i had to write it to be able
to patch entries.

Also lots of visual changes and improvements, added a color gradient
algorithm for visualization.

This code is in need of some serious refactoring.