Started FE 2021 rewrite
This rewrite will require some changes to the underlying structure of
the app, including relying on Vue-router for managing multiple views.

Started working on these changes, with first establishing the router
structure, and the all-important conditional routing "loading" screen
which will help users by figuring out where to go.

This commit is just a skeleton-commit, not functional yet.

(also, this rewrite is a great opportunity for rewriting some stuff that
I've been dying to fix!!)